# api/openapi.yaml
openapi: 3.1.0
info:
  title: OnboardMePlease API
  version: 0.4.0
  description: >-
    Capture, resumable synthesis, searchable OKF knowledge and source-grounded questions are implemented.
    Deletion and separate snapshot creation remain planned; those routes are marked below.
servers:
  - url: http://127.0.0.1:8765
    description: Loopback default for the local single-user service.
paths:
...
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/index:
    post:
      operationId: indexSnapshot
      summary: Reindex an existing ready GitHub snapshot
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '202': { description: Durable indexing job queued }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/search:
    get:
      operationId: searchEvidence
      summary: Scoped exact/lexical or explicitly selected semantic retrieval
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: mode
          in: query
          schema: { enum: [lexical, semantic] }
        - name: provider
          in: query
          schema: { enum: [local, cloud] }
      responses:
        '200': { description: Snapshot-scoped source chunks with ranges }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/capabilities:
    get:
      operationId: getCapabilities
      summary: Per-artifact analysis capabilities and gaps
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Artifact capability records }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/symbols:
    get:
      operationId: getSymbols
      summary: Snapshot-scoped extracted or SCIP-imported symbols
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200': { description: Symbols and source ranges }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/graph:
    get:
      operationId: getGraph
      summary: Bounded typed relation traversal from one symbol
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: symbol_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Nodes, edges, limits, and truncation flag }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/scip:
    post:
      operationId: importSCIP
      summary: Import matching SCIP semantic evidence
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200': { description: Imported and skipped document counts }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/embeddings:
    post:
      operationId: queueEmbeddings
      summary: Explicitly queue local or opted-in cloud embeddings
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [provider]
              properties:
                provider: { enum: [local, cloud] }
      responses:
        '202': { description: Durable embedding job queued }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/evidence/{evidenceId}:
    get:
      operationId: getEvidence
      summary: Resolve an authorized citation to sanitized source lines
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: evidenceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sanitized evidence; never raw excluded secret content
          content:
            application/json:
              schema: { $ref: '../contracts/evidence.schema.json#/$defs/Evidence' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
  responses:
    Problem:
      description: Safe error without raw input, code, path, or credential value
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }